requirements.txt
jikanpy ==  4.2.2
discord.py == 1.4.1
python-dotenv == 0.14.0